Rename files to comply with standards
